refactor(server): creating QBCore.Functions in a module
