+++ fxmanifest.lua
@@ -29,7 +29,6 @@ server_scripts {
     'server/player.lua',
     'server/events.lua',
     'server/commands.lua',
-    'server/exports.lua',
     'server/debug.lua',
     'server/loops.lua',
     'server/storage.lua',
