+++ server/main.lua
@@ -8,6 +8,14 @@ QBCore = {}
 QBCore.Config = QBConfig
 QBCore.Shared = QBShared
 
+QBCore.Player_Buckets = {}
+QBCore.Entity_Buckets = {}
+QBCore.UsableItems = {}
+QBCore.Functions = require 'functions'
+
+---@deprecated Call lib.print.debug() instead
+QBCore.Debug = DebugPrint
+
 ---@deprecated use https://overextended.github.io/docs/ox_lib/Callback/Lua/Server instead
 QBCore.ClientCallbacks = {}
